Re-created 2.4 as 2.4.2
I was unable to un-compile my previos code, but this is pretty much the same as before.

1. Declare Variables
2. print variables in reverse order

one thing that confuses me is the "%i" or whatever the variable might entail, becuase in Java, all you do is System.out.print(<variable>); so this language seems to be a more complicated syntax of Java... which makes sense but I'm not quiet sure what that syntax is.